Adding performance tests.
Performance tests are generated as a part of Unity.Mathematics code
generation and are only meant to be run locally using the included
test project.  They are not run as a part of CI in the GitHub
repository.

The performance tests are set up to run both burst and mono versions
of a given piece of code so you can track the performance of each
separately.