Added Random.CreateFromIndex(). (#128)
Random.CreateFromHashedSeed() is intended to be used when multiple
instances of Random need to be created from consecutive seeds.  This
is common when the seed is derived from an index of an array.  The
hashing helps to ensure that the Random state is very different for
each index and reduces the chance that the numbers drawn from those
Random instances are similar.

Fixes #78