Separate the getter and setter tests for swizzles.
This required me to create a new class to accumulate the code gen into
two separate StringBuilders, one for getter and one for setter. Doing
this allows me to iterate over all the permutations just once.

Some things still need to be cleaned up like the delegate signature
which assumes it takes a StringBuilder to append to, but the new class
doesn't use that. Also needs to support two bool vectors as sources.