Fix test asserts that fail in Tiny IL2CPP builds.
https://unity3d.atlassian.net/browse/DOTS-1745

NaN and signed zeros were not matching the expected bit pattern so the
tests were changed to explicitly use the correct value.

Additionally, the pow() function in il2cpp appears to have a bug where
it handles a base of negative infinity incorrectly. The asserts
corresponding to those cases were removed while a fix to il2cpp is
made. This needs a follow-up: https://unity3d.atlassian.net/browse/DOTS-1833