Make performance tests create a variable for iteration count. (#205)
While the performance tests worked just fine, it was annoying to hand
modify a performance test when investigating performance issues. You
had to search the code for all instances of a const literal and be
sure you changed the number in all those places. Instead, use a
variable so you can change it in one place.