Make tests run on Android player builds.
There is some strange behavior going on with TestCompiler and TestCase
that may need to be followed up on. When only using TestCompiler, it
appeared that those tests did not get run in the Android player build.
But now that I have both TestCompiler AND TestCase, all the tests seem
to run.